requirements.txt
openai>=1.50.0
pydantic>=2.7.0
PyMuPDF>=1.24.1
pytesseract>=0.3.10
Pillow>=10.2.0
PyYAML>=6.0.1
pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.1
pypdf
requests
google-generativeai